feat(app): enhance SEO experience in home page and product page
