+++ src/app/page.tsx
@@ -1,9 +1,19 @@
-import HomePage from "./home/page";
+import HomePageContent from "./home/section/content";
+
+export async function generateMetadata() {
+  return {
+    title: 'Tyeso Indonesia',
+    description: 'Discover our newest collection—where style meets functionality in every piece. Fresh designs, vibrant colors, and premium quality await to elevate your everyday essentials!',
+    keywords: ['Tyeso', 'Product', 'Bottle']
+  }
+}
 
 export default function Home() {
   return (
     <main className="min-h-screen min-w-scren">
-      <HomePage />
+      <div className="w-screen h-fit p-0 m-0 bg-white">
+        <HomePageContent />
+      </div>
     </main>
   );
 }
